Remove IComplexityAnalyzer
