+++ src/GraphQL.Tests/DI/GraphQLBuilderExtensionTests.cs
@@ -340,19 +340,10 @@ public class GraphQLBuilderExtensionTests
 
     #region - AddComplexityAnalyzer -
     [Theory]
-    [InlineData(false, false, false, false, false)]
-    [InlineData(false, false, false, true, false)]
-    [InlineData(false, false, false, false, true)]
-    [InlineData(true, false, false, false, false)]
-    [InlineData(true, false, false, true, false)]
-    [InlineData(true, false, false, false, true)]
-    [InlineData(true, true, false, false, false)]
-    [InlineData(true, true, false, true, false)]
-    [InlineData(true, true, false, false, true)]
-    [InlineData(true, false, true, false, false)]
-    [InlineData(true, false, true, true, false)]
-    [InlineData(true, false, true, false, true)]
-    public void AddComplexityAnalyzer(bool typed, bool withFactory, bool withInstance, bool withAction, bool withAction2)
+    [InlineData(false, false)]
+    [InlineData(true, false)]
+    [InlineData(false, true)]
+    public void AddComplexityAnalyzer(bool withAction, bool withAction2)
     {
         var ruleInstance = new ComplexityValidationRule(new ComplexityConfiguration());
         MockSetupRegister<ComplexityValidationRule, ComplexityValidationRule>();
@@ -360,59 +351,17 @@ public class GraphQLBuilderExtensionTests
         var mockServiceProvider = new Mock<IServiceProvider>(MockBehavior.Strict);
         mockServiceProvider.Setup(s => s.GetService(typeof(ComplexityValidationRule))).Returns(ruleInstance).Verifiable();
         var getOpts = MockSetupConfigureExecution(mockServiceProvider.Object);
-        Func<IServiceProvider, IComplexityAnalyzer> factory = null!;
-        MyComplexityAnalyzer instance = null!;
-        if (typed)
-        {
-            if (withFactory)
-            {
-                factory = MockSetupRegister<IComplexityAnalyzer>();
-            }
-            else if (withInstance)
-            {
-                instance = new MyComplexityAnalyzer();
-                MockSetupRegister<IComplexityAnalyzer>(instance);
-            }
-            else
-            {
-                MockSetupRegister<IComplexityAnalyzer, MyComplexityAnalyzer>();
-            }
-        }
         if (withAction)
         {
             var action = MockSetupConfigure1<ComplexityConfiguration>();
-            if (typed)
-            {
-                if (withFactory)
-                    _builder.AddComplexityAnalyzer(factory, action);
-                else if (withInstance)
-                    _builder.AddComplexityAnalyzer(instance, action);
-                else
-                    _builder.AddComplexityAnalyzer<MyComplexityAnalyzer>(action);
-            }
-            else
-            {
-                _builder.AddComplexityAnalyzer(action);
-            }
+            _builder.AddComplexityAnalyzer(action);
         }
         else
         {
             var action = withAction2 ? MockSetupConfigure2<ComplexityConfiguration>() : null;
             if (!withAction2)
                 MockSetupConfigureNull<ComplexityConfiguration>();
-            if (typed)
-            {
-                if (withFactory)
-                    _builder.AddComplexityAnalyzer(factory, action!);
-                else if (withInstance)
-                    _builder.AddComplexityAnalyzer(instance, action!);
-                else
-                    _builder.AddComplexityAnalyzer<MyComplexityAnalyzer>(action!);
-            }
-            else
-            {
-                _builder.AddComplexityAnalyzer(action!);
-            }
+            _builder.AddComplexityAnalyzer(action!);
         }
         var opts = getOpts();
         opts.ValidationRules.ShouldNotBeNull();
@@ -421,15 +370,6 @@ public class GraphQLBuilderExtensionTests
         mockServiceProvider.Verify();
         Verify();
     }
-
-    [Fact]
-    public void AddComplexityAnalyzer_Null()
-    {
-        Should.Throw<ArgumentNullException>(() => _builder.AddComplexityAnalyzer((IComplexityAnalyzer)null!));
-        Should.Throw<ArgumentNullException>(() => _builder.AddComplexityAnalyzer((Func<IServiceProvider, IComplexityAnalyzer>)null!));
-    }
-
-    private class MyComplexityAnalyzer : ComplexityAnalyzer { }
     #endregion
 
     #region - AddErrorInfoProvider -
@@ -1348,8 +1288,11 @@ public class GraphQLBuilderExtensionTests
     {
     }
 
-    private class TestComplexityAnalyzer : ComplexityAnalyzer
+    private class TestComplexityAnalyzer : ComplexityValidationRule
     {
+        public TestComplexityAnalyzer() : base(new())
+        {
+        }
     }
 
     private class TestErrorInfoProvider : ErrorInfoProvider
