+++ src/GraphQL/Extensions/GraphQLBuilderExtensions.cs
@@ -398,81 +398,6 @@ public static class GraphQLBuilderExtensions // TODO: split
         builder.Services.Configure(action);
         return builder;
     }
-
-    /// <summary>
-    /// Registers <typeparamref name="TAnalyzer"/> as a singleton of type <see cref="IComplexityAnalyzer"/> within the
-    /// dependency injection framework, then enables and configures it with the specified configuration delegate.
-    /// </summary>
-    [Obsolete("Please write a custom complexity analyzer as a validation rule. This method will be removed in v8.")]
-    public static IGraphQLBuilder AddComplexityAnalyzer<TAnalyzer>(this IGraphQLBuilder builder, Action<ComplexityConfiguration>? action = null)
-        where TAnalyzer : class, IComplexityAnalyzer
-    {
-        builder.Services.Register<IComplexityAnalyzer, TAnalyzer>(ServiceLifetime.Singleton);
-        builder.AddValidationRule<ComplexityValidationRule>();
-        builder.Services.Configure(action);
-        return builder;
-    }
-
-    /// <inheritdoc cref="AddComplexityAnalyzer{TAnalyzer}(IGraphQLBuilder, Action{ComplexityConfiguration})"/>
-    [Obsolete("Please write a custom complexity analyzer as a validation rule. This method will be removed in v8.")]
-    public static IGraphQLBuilder AddComplexityAnalyzer<TAnalyzer>(this IGraphQLBuilder builder, Action<ComplexityConfiguration, IServiceProvider?>? action)
-        where TAnalyzer : class, IComplexityAnalyzer
-    {
-        builder.Services.Register<IComplexityAnalyzer, TAnalyzer>(ServiceLifetime.Singleton);
-        builder.AddValidationRule<ComplexityValidationRule>();
-        builder.Services.Configure(action);
-        return builder;
-    }
-
-    /// <summary>
-    /// Registers <paramref name="analyzer"/> as a singleton of type <see cref="IComplexityAnalyzer"/> within the
-    /// dependency injection framework, then enables and configures it with the specified configuration delegate.
-    /// </summary>
-    [Obsolete("Please write a custom complexity analyzer as a validation rule. This method will be removed in v8.")]
-    public static IGraphQLBuilder AddComplexityAnalyzer<TAnalyzer>(this IGraphQLBuilder builder, TAnalyzer analyzer, Action<ComplexityConfiguration>? action = null)
-        where TAnalyzer : class, IComplexityAnalyzer
-    {
-        builder.Services.Register<IComplexityAnalyzer>(analyzer ?? throw new ArgumentNullException(nameof(analyzer)));
-        builder.AddValidationRule<ComplexityValidationRule>();
-        builder.Services.Configure(action);
-        return builder;
-    }
-
-    /// <inheritdoc cref="AddComplexityAnalyzer{TAnalyzer}(IGraphQLBuilder, TAnalyzer, Action{ComplexityConfiguration})"/>
-    [Obsolete("Please write a custom complexity analyzer as a validation rule. This method will be removed in v8.")]
-    public static IGraphQLBuilder AddComplexityAnalyzer<TAnalyzer>(this IGraphQLBuilder builder, TAnalyzer analyzer, Action<ComplexityConfiguration, IServiceProvider?>? action)
-        where TAnalyzer : class, IComplexityAnalyzer
-    {
-        builder.Services.Register<IComplexityAnalyzer>(analyzer ?? throw new ArgumentNullException(nameof(analyzer)));
-        builder.AddValidationRule<ComplexityValidationRule>();
-        builder.Services.Configure(action);
-        return builder;
-    }
-
-    /// <summary>
-    /// Registers a singleton of type <see cref="IComplexityAnalyzer"/> within the dependency injection framework
-    /// using the specified factory delegate, then enables and configures it with the specified configuration delegate.
-    /// </summary>
-    [Obsolete("Please write a custom complexity analyzer as a validation rule. This method will be removed in v8.")]
-    public static IGraphQLBuilder AddComplexityAnalyzer<TAnalyzer>(this IGraphQLBuilder builder, Func<IServiceProvider, TAnalyzer> analyzerFactory, Action<ComplexityConfiguration>? action = null)
-        where TAnalyzer : class, IComplexityAnalyzer
-    {
-        builder.Services.Register<IComplexityAnalyzer>(analyzerFactory ?? throw new ArgumentNullException(nameof(analyzerFactory)), ServiceLifetime.Singleton);
-        builder.AddValidationRule<ComplexityValidationRule>();
-        builder.Services.Configure(action);
-        return builder;
-    }
-
-    /// <inheritdoc cref="AddComplexityAnalyzer{TAnalyzer}(IGraphQLBuilder, Func{IServiceProvider, TAnalyzer}, Action{ComplexityConfiguration})"/>
-    [Obsolete("Please write a custom complexity analyzer as a validation rule. This method will be removed in v8.")]
-    public static IGraphQLBuilder AddComplexityAnalyzer<TAnalyzer>(this IGraphQLBuilder builder, Func<IServiceProvider, TAnalyzer> analyzerFactory, Action<ComplexityConfiguration, IServiceProvider?>? action)
-        where TAnalyzer : class, IComplexityAnalyzer
-    {
-        builder.Services.Register<IComplexityAnalyzer>(analyzerFactory ?? throw new ArgumentNullException(nameof(analyzerFactory)), ServiceLifetime.Singleton);
-        builder.AddValidationRule<ComplexityValidationRule>();
-        builder.Services.Configure(action);
-        return builder;
-    }
     #endregion
 
     #region - AddErrorInfoProvider
