+++ src/GraphQL/Validation/Rules.Custom/ComplexityValidationRule.cs
@@ -1,5 +1,8 @@
+using System.Diagnostics;
+using GraphQL.Types;
 using GraphQL.Validation.Complexity;
 using GraphQL.Validation.Errors.Custom;
+using GraphQLParser;
 using GraphQLParser.AST;
 
 namespace GraphQL.Validation.Rules.Custom;
@@ -7,78 +10,198 @@ namespace GraphQL.Validation.Rules.Custom;
 /// <summary>
 /// Analyzes a document to determine if its complexity exceeds a threshold.
 /// </summary>
-public class ComplexityValidationRule : IValidationRule
+public class ComplexityValidationRule : IValidationRule, INodeVisitor
 {
     private readonly ComplexityConfiguration _complexityConfiguration;
-#pragma warning disable CS0618 // Type or member is obsolete
-    private readonly IComplexityAnalyzer _complexityAnalyzer;
-#pragma warning restore CS0618 // Type or member is obsolete
 
     /// <summary>
     /// Initializes an instance with the specified complexity configuration.
     /// </summary>
     public ComplexityValidationRule(ComplexityConfiguration complexityConfiguration)
-#pragma warning disable CS0618 // Type or member is obsolete
-        : this(complexityConfiguration, new ComplexityAnalyzer())
-#pragma warning restore CS0618 // Type or member is obsolete
-    {
-    }
-
-    /// <summary>
-    /// Initializes an instance with the specified complexity configuration and complexity analyzer.
-    /// </summary>
-    [Obsolete("Please write a custom complexity analyzer as a validation rule. This constructor will be removed in v8.")]
-    public ComplexityValidationRule(ComplexityConfiguration complexityConfiguration, IComplexityAnalyzer complexityAnalyzer)
     {
         _complexityConfiguration = complexityConfiguration;
-        _complexityAnalyzer = complexityAnalyzer;
     }
 
     /// <inheritdoc/>
-    public ValueTask<INodeVisitor?> ValidateAsync(ValidationContext context)
-        => new(new PostponeComplexityValidationVisitor(_complexityConfiguration, _complexityAnalyzer));
+    public ValueTask<INodeVisitor?> ValidateAsync(ValidationContext context) => new(this);
 
-    // https://github.com/graphql-dotnet/graphql-dotnet/issues/3527
-    private sealed class PostponeComplexityValidationVisitor : INodeVisitor
+    ValueTask INodeVisitor.EnterAsync(ASTNode node, ValidationContext context) => default;
+
+    ValueTask INodeVisitor.LeaveAsync(ASTNode node, ValidationContext context)
     {
-        private readonly ComplexityConfiguration _complexityConfiguration;
-#pragma warning disable CS0618 // Type or member is obsolete
-        private readonly IComplexityAnalyzer _complexityAnalyzer;
-#pragma warning restore CS0618 // Type or member is obsolete
-
-#pragma warning disable CS0618 // Type or member is obsolete
-        public PostponeComplexityValidationVisitor(ComplexityConfiguration complexityConfiguration, IComplexityAnalyzer complexityAnalyzer)
+        // Complexity analysis should run at the very end.
+        if (node is GraphQLDocument)
         {
-            _complexityConfiguration = complexityConfiguration;
-            _complexityAnalyzer = complexityAnalyzer;
-        }
-#pragma warning restore CS0618 // Type or member is obsolete
+            // Fast return here to avoid all possible problems with complexity analysis.
+            // For example, document may contain fragment cycles, see https://github.com/graphql-dotnet/graphql-dotnet/issues/3527
+            // Note, that ComplexityValidationRule/ComplexityAnalyzer still checks for fragment cycles since there may be no standard rules configured.
+            if (context.HasErrors)
+                return default;
 
-        public ValueTask EnterAsync(ASTNode node, ValidationContext context) => default;
-
-        public ValueTask LeaveAsync(ASTNode node, ValidationContext context)
-        {
-            // Complexity analysis should run at the very end.
-            if (node is GraphQLDocument)
+            try
             {
-                // Fast return here to avoid all possible problems with complexity analysis.
-                // For example, document may contain fragment cycles, see https://github.com/graphql-dotnet/graphql-dotnet/issues/3527
-                // Note, that ComplexityValidationRule/ComplexityAnalyzer still checks for fragment cycles since there may be no standard rules configured.
-                if (context.HasErrors)
-                    return default;
+                using (context.Metrics.Subject("document", "Analyzing complexity"))
+                    Validate(context.Document, _complexityConfiguration, context.Schema);
+            }
+            catch (ComplexityError ex)
+            {
+                context.ReportError(ex);
+            }
+        }
 
-                try
+        return default;
+    }
+
+    /// <inheritdoc/>
+    public void Validate(GraphQLDocument document, ComplexityConfiguration complexityParameters, ISchema schema)
+    {
+        if (complexityParameters == null)
+            return;
+        var complexityResult = Analyze(document, complexityParameters.FieldImpact ?? 2.0f, complexityParameters.MaxRecursionCount, schema);
+
+        Analyzed(document, complexityParameters, complexityResult);
+
+        if (complexityResult.Complexity > complexityParameters.MaxComplexity)
+            throw new ComplexityError(
+                $"Query is too complex to execute. Complexity is {complexityResult.Complexity}, maximum allowed on this endpoint is {complexityParameters.MaxComplexity}. The field with the highest complexity is '{GetName(complexityResult.ComplexityMap.OrderByDescending(pair => pair.Value).First().Key)}' with value {complexityResult.ComplexityMap.OrderByDescending(pair => pair.Value).First().Value}.");
+
+        if (complexityResult.TotalQueryDepth > complexityParameters.MaxDepth)
+            throw new ComplexityError(
+                $"Query is too nested to execute. Depth is {complexityResult.TotalQueryDepth} levels, maximum allowed on this endpoint is {complexityParameters.MaxDepth}.");
+
+        string GetName(ASTNode node)
+        {
+            return node switch
+            {
+                GraphQLField f => f.Name.StringValue,
+                GraphQLFragmentSpread fs => fs.FragmentName.Name.StringValue,
+                _ => throw new NotSupportedException(node.ToString()),
+            };
+        }
+    }
+
+    /// <summary>
+    /// Executes after the complexity analysis has completed, before comparing results to the complexity configuration parameters.
+    /// This method is made to be able to access the calculated <see cref="ComplexityResult"/> and handle it, for example, for logging.
+    /// </summary>
+    protected virtual void Analyzed(GraphQLDocument document, ComplexityConfiguration complexityParameters, ComplexityResult complexityResult)
+    {
+#if DEBUG
+        Debug.WriteLine($"Complexity: {complexityResult.Complexity}");
+        Debug.WriteLine($"Sum(Query depth across all subqueries) : {complexityResult.TotalQueryDepth}");
+        foreach (var node in complexityResult.ComplexityMap)
+            Debug.WriteLine($"{node.Key} : {node.Value}");
+#endif
+    }
+
+    /// <summary>
+    /// Analyzes the complexity of a document.
+    /// </summary>
+    internal ComplexityResult Analyze(GraphQLDocument doc, double avgImpact, int maxRecursionCount, ISchema? schema = null)
+    {
+        if (avgImpact <= 1)
+            throw new ArgumentOutOfRangeException(nameof(avgImpact));
+
+        var context = new AnalysisContext
+        {
+            MaxRecursionCount = maxRecursionCount,
+            AvgImpact = avgImpact,
+            CurrentEndNodeImpact = 1d
+        };
+        var visitor = schema == null ? new ComplexityVisitor() : new ComplexityVisitor(schema);
+
+        // https://github.com/graphql-dotnet/graphql-dotnet/issues/3030
+        // Sort fragment definitions so that independent fragments go in front.
+        var dependencies = BuildDependencies(doc);
+        List<GraphQLFragmentDefinition> orderedFragments = new();
+        List<GraphQLFragmentDefinition> fragsToNull = new();
+
+        while (dependencies.Count > 0)
+        {
+            var independentFragment = GetFirstFragmentWithoutPendingDependencies(dependencies);
+
+            orderedFragments.Add(independentFragment);
+            dependencies.Remove(independentFragment);
+
+            foreach (var item in dependencies) // no deconstruct syntax for netstandard2.0
+            {
+                if (item.Value?.Remove(independentFragment) == true && item.Value.Count == 0)
+                    fragsToNull.Add(item.Key);
+            }
+
+            // next candidates for GetFirstFragmentWithoutPendingDependencies
+            foreach (var frag in fragsToNull)
+                dependencies[frag] = null;
+
+            fragsToNull.Clear();
+        }
+
+        foreach (var frag in orderedFragments)
+            visitor.VisitAsync(frag, context).GetAwaiter().GetResult();
+
+        context.FragmentMapAlreadyBuilt = true;
+
+        visitor.VisitAsync(doc, context).GetAwaiter().GetResult();
+
+        return context.Result;
+    }
+
+    private static GraphQLFragmentDefinition GetFirstFragmentWithoutPendingDependencies(Dictionary<GraphQLFragmentDefinition, HashSet<GraphQLFragmentDefinition>?> dependencies)
+    {
+        foreach (var item in dependencies)
+        {
+            if (item.Value == null)
+                return item.Key;
+        }
+
+        throw new InvalidOperationException("Fragments dependency cycle detected!");
+    }
+
+    private static Dictionary<GraphQLFragmentDefinition, HashSet<GraphQLFragmentDefinition>?> BuildDependencies(GraphQLDocument document)
+    {
+        Stack<GraphQLSelectionSet> _selectionSetsToVisit = new();
+        Dictionary<GraphQLFragmentDefinition, HashSet<GraphQLFragmentDefinition>?> dependencies = new();
+
+        foreach (var fragmentDef in document.Definitions.OfType<GraphQLFragmentDefinition>())
+        {
+            dependencies[fragmentDef] = GetDependencies(fragmentDef);
+        }
+
+        return dependencies;
+
+        HashSet<GraphQLFragmentDefinition>? GetDependencies(GraphQLFragmentDefinition def)
+        {
+            HashSet<GraphQLFragmentDefinition>? dependencies = null;
+            _selectionSetsToVisit.Push(def.SelectionSet);
+
+            int counter = 0;
+            const int MAX_ITERATIONS = 2000;
+            while (_selectionSetsToVisit.Count > 0)
+            {
+                // https://github.com/graphql-dotnet/graphql-dotnet/issues/3527
+                if (++counter > MAX_ITERATIONS)
+                    throw new ValidationError("It looks like document has fragment cycle. Please make sure you are using standard validation rules especially NoFragmentCycles one.");
+
+                foreach (var selection in _selectionSetsToVisit.Pop().Selections)
                 {
-                    using (context.Metrics.Subject("document", "Analyzing complexity"))
-                        _complexityAnalyzer.Validate(context.Document, _complexityConfiguration, context.Schema);
-                }
-                catch (ComplexityError ex)
-                {
-                    context.ReportError(ex);
+                    if (selection is GraphQLFragmentSpread spread)
+                    {
+                        var frag = document.FindFragmentDefinition(spread.FragmentName.Name.Value);
+                        if (frag != null)
+                        {
+                            (dependencies ??= new()).Add(frag);
+                            _selectionSetsToVisit.Push(frag.SelectionSet);
+                        }
+                    }
+                    else if (selection is IHasSelectionSetNode hasSet && hasSet.SelectionSet != null)
+                    {
+                        _selectionSetsToVisit.Push(hasSet.SelectionSet);
+                    }
                 }
             }
 
-            return default;
+            _selectionSetsToVisit.Clear();
+            return dependencies;
         }
     }
 }
